//월별,연별 그래프 완성2//쿼리만 추가하면 됨.
//검색 콤보 박스 완성//검색 수정할 차례
//검색기능 추가 해야함
//새 표에 들어갈 내용 파악 해야함
//콤보박스 중복 안되는 문제
//총 주문건수

+++ src/main/java/kr/happyjob/study/busOdm/model/OrderModel.java
@@ -13,7 +13,7 @@ public class OrderModel {
 	private String clnt_add_dt;  // 고객기업상세주소
 	private String loginID;  // 사번
 	private String name;  // 영업담당자(이름)
-	private int order_tot_price;  // 총주문금액
+	private long order_tot_price;  // 총주문금액
 	private String order_cancel;  // 취소여부
 	private String order_date;  // 주문날짜
 	private String order_req;  // 요청사항
@@ -22,7 +22,7 @@ public class OrderModel {
 	private int product_no;  // 제품번호
 	private int product_price;  // 판매가
 	private int order_dt_amt;  // 주문수량
-	private int order_dt_price;  // 주문금액
+	private long order_dt_price;  // 주문금액
 	
 	// 제품정보 테이블 (tb_product)
 	private String product_name;  // 제품명
@@ -58,10 +58,10 @@ public class OrderModel {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getOrder_tot_price() {
+	public long getOrder_tot_price() {
 		return order_tot_price;
 	}
-	public void setOrder_tot_price(int order_tot_price) {
+	public void setOrder_tot_price(long order_tot_price) {
 		this.order_tot_price = order_tot_price;
 	}
 	public String getOrder_cancel() {
@@ -142,10 +142,10 @@ public class OrderModel {
 	public void setDetail_price(int detail_price) {
 		this.detail_price = detail_price;
 	}
-	public int getOrder_dt_price() {
+	public long getOrder_dt_price() {
 		return order_dt_price;
 	}
-	public void setOrder_dt_price(int order_dt_price) {
+	public void setOrder_dt_price(long order_dt_price) {
 		this.order_dt_price = order_dt_price;
 	}
